Fix resumes.astro
Enabled SSR?

Found a way to get querystring! This is amazing.

Changed the way src/pages/resumes.astro detects which resumes there are so that npm run build doesn't run into errors. For some reason .astro files can't use any node libraries. It's pretty stupid, or so my small brain thinks.

I tried to fix up visualization today but failed. There's a now known issue where each dataset is sorted... but that means that the wrong people have the wrong data associated with them as a result of this sorting. I need to fix this!

Change node js to the website's middleware. IDK what that does honestly. I think I did it for the SSR.