Chart + Slightly CSS Optimization
In daily.json, formatted the header as an array of strings, not an array with one array of strings.

Changed the css files slightly to hopefully get slightly better runtime. Color variables are now applied to the HTML document element rather than EVERY SINGLE ELEMENT IN THE DOCUMENT.

Added a basic chart to the Data webpage. I realize now that I probably will not be able to generalize charts (AKA I can't have a single component to take care of everything).