Content Description, Directives page, and Image Centering
So it appears that anything inside the body of example-image.md & example-text.md does not appear inside the awards section. This is intended behaviour, but the example inside the content folder implied that it wasn't. Now it does imply that it's intended!

Added directives page.

Images for awards & directives now center. Yay!