Graphing Regression
The regression line is now graphed. It is a known bug that the regression line can travel outside of the chart, but in the context of my personal data, it will probably never do that.

If I want to make this into a full-on Chart js plugin, I'll have to solve that big problem. But horray, because I solved a big issue I had with the charts!