Quick fixes
The font for the top-bar did not match the rest of the webpage.

Also, the margin for the body of schedule.astro was not large enough. It is fixed now, by hardcoding (AKA this code might break in the future).

AGAIN, an md file implied that the text inside is the actual content. I fixed that now.