Add dark mode Wywy falling background
So JPG has an issue of not having the option to be transparent. The issue is that if the jpg image goes behind a Card, a person can easily see the giant SQUARE of color :P. This is why I opted to use png files, even though their compression kinda sucks. Maybe I should downgrade the resolution of the image later on.