Slight Improvements
Added empty alt attributes for the decorative falling background.

Added alt attribute for the website icon

Changed names of elements to be more... semantic? I forgot the word for it (i.e. div -> nav)

Ensured that the minimum height for the page is 100vh, not 100vh + top bar height.

Removed the height restriction for the footer. It should now default to a 75px size and grow if it needs to.