BUGGED - Baseline code for coldTurkeyBlocker.py
coldTurkeyBlocker.py is intended to help fetch screentime data.

I eventually plan on consolidating this to a "daily script" that runs via Astro or Node or Next or Nuxt locally and makes the computer go around, look for data, and push it to a sheet so that recording data is super easy. But I'm a long ways from being there!