[BROKEN] Change Validation & Form Output
Corrected a mistake in app.c where it would account for the null terminator when the match size does not include a null terminator.

Separated the values for the zod schema and the output POST query to the database to make it easier to code the sql-receptionist (most data conversion should happen client-side).