Docker (Debian) w. C
Stopped using gcc:latest image, which caused problems with packages, and opted for Debian Bookworm instead.

Changed the makefile to prevent it from running the app & hanging RUN inside the Docker.

Also changed the file extension of the makefile's output to be consistent with Linux conventions (.exe is usually used on Windows)

Added a helper script to completely purge all Docker images & containers.